refactor(threads): useThreadActions 抽 indexEarlyPaint 与 engineAsyncMerge 两段
- 3066→2244 行：indexEarlyPaint 324（Session Index 早画）+ engineAsyncMerge 815（六引擎异步 merge）
- dispatch 次序与 ownsRequest 结算（L2956-2966 语义）原位逐字；deps 逐字
- 主文件 <2400 满足 M2 feature-hotpath 门槛
- baseline 重生成 + 新文件登记豁免
验证：tsc 全仓零 error；25 测试文件 205 passed/77 failed 与同树基线逐一相同；gate 绿
（子代理 ThreadActionsSplit 执行，主代理复核验证）
