+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T20:42:22.381Z",
+  "generatedAt": "2026-08-30T20:51:01.210Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
